Add libGDX dep, use ClassReflection.
This adds the collision and geometry packages to the class reflection pool, which is probably overkill. It also switches formic to use the more-reliably-available Maven Central dependency (without 'v') and updates jUnit to the recommended version 4.13.2 (it might have issues if running on JDK 7, but 4.13.1 also has some type of security issue on JDK 7).